move silly km to xor km.  Add delay reply to node.
