+++ adapter/ph/src/lib.rs
@@ -15,9 +15,10 @@ pub mod defs;
 pub mod fastpath;
 pub mod flow_control;
 pub mod km;
+pub mod km_noise;
+pub mod km_xor;
 pub mod mgmt;
 pub mod mgmt_processor_worker;
-pub mod km_noise;
 pub mod net_defs;
 pub mod options;
 pub mod packet;
